Revert "Backend test"
