--- ProjectBackEnd/Gamification/Gamification/AutoMappingProfile.cs
@@ -1,19 +1,0 @@
-﻿using AutoMapper;
-using Gamification.BLL.DTO;
-using Gamification.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Gamification
-{
-    public class AutoMappingProfile : Profile
-    {
-        public AutoMappingProfile()
-        {
-            CreateMap<Achievement, AchievementDTO>();
-            CreateMap<AchievementDTO, Achievement>();
-        }
-    }
-}
