--- ProjectBackEnd/Gamification/Gamification/BLL/DTO/AchievementDTO.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Gamification.BLL.DTO
-{
-    public class AchievementDTO
-    {
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public int Xp { get; set; }
-        public int IconId { get; set; }
-    }
-}
