+++ ProjectBackEnd/Gamification/Gamification/Controllers/AchievementController.cs
@@ -9,9 +9,6 @@ using Microsoft.Extensions.Logging;
 using System.Threading;
 using Microsoft.EntityFrameworkCore;
 using Gamification.DAL.IRepository;
-using AutoMapper;
-using Gamification.BLL.DTO;
-using Gamification.DAL.Repository.UnitOfWork;
 
 namespace Gamification.Controllers
 {
@@ -19,54 +16,24 @@ namespace Gamification.Controllers
     [ApiController]
     public class AchievementController : ControllerBase
     {
-        private readonly IMapper _mapper;
-        private IUnitOfWork _unitOfWork { get; set; }
-        public AchievementController(IUnitOfWork unitOfWork, IMapper mapper)
+        private readonly IAchievementRepository _achievementRepository;
+        public AchievementController(IAchievementRepository achievementRepository)
         {
-            _unitOfWork = unitOfWork;
-            _mapper = mapper;
+            _achievementRepository = achievementRepository;
         }
 
         [HttpGet]
         public async Task<IActionResult> GetAllAchievements(CancellationToken cancellationToken)
         {
-            var achievements = await _unitOfWork.achievementRepository.GetAllAchievements(cancellationToken);
-            var mapData = _mapper.Map<IEnumerable<AchievementDTO>>(achievements);
-            return Ok(mapData);
+            var achievements = _achievementRepository.GetAllAchievements(cancellationToken);
+            return Ok(await achievements);
         }
 
         [HttpGet("{Id}")]
-        public async Task<IActionResult> GetAchievementById(Guid Id, CancellationToken cancellationToken)
+        public async Task<ActionResult<Achievement>> GetAchievementById(int Id, CancellationToken cancellationToken)
         {
-            var achievement = await _unitOfWork.achievementRepository.GetAchievementById(Id, cancellationToken);
-            var mapData = _mapper.Map<AchievementDTO>(achievement);
-            return Ok(mapData);
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> CreateAchievement(AchievementDTO newAchievement, CancellationToken cancellationToken)
-        {
-            var mapData = _mapper.Map<Achievement>(newAchievement);
-            var achievement = _unitOfWork.achievementRepository.CreateAchievement(mapData, cancellationToken);
-            await _unitOfWork.SaveChanges(cancellationToken);
-            return Ok(achievement);
-        }
-
-        [HttpPut]
-        public async Task<IActionResult> UpdateAchievement(AchievementDTO newAchievement, CancellationToken cancellationToken)
-        {
-            var mapData = _mapper.Map<Achievement>(newAchievement);
-            var achievement = _unitOfWork.achievementRepository.UpdateAchievement(mapData, cancellationToken);
-            await _unitOfWork.SaveChanges(cancellationToken);
-            return Ok(achievement);
-        }
-
-        [HttpDelete]
-        public async Task<IActionResult> DeleteAchievement(Guid achievemenId, CancellationToken cancellationToken)
-        {
-            var deletedAcievement = _unitOfWork.achievementRepository.DeleteAchievement(achievemenId, cancellationToken);
-            await _unitOfWork.SaveChanges(cancellationToken);
-            return Ok(deletedAcievement);
+            var achievement = _achievementRepository.GetAchievementById(Id, cancellationToken);
+            return Ok(await achievement);
         }
     }
 }
