--- ProjectBackEnd/Gamification/Gamification/DAL/IRepositories/IAchievementRepository.cs
@@ -1,19 +1,0 @@
-﻿using Gamification.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Gamification.DAL.IRepository
-{
-    public interface IAchievementRepository
-    {
-        public Task<IEnumerable<Achievement>> GetAllAchievements(CancellationToken cancellationToken);
-        public Task<Achievement> GetAchievementById(Guid AchievementId, CancellationToken cancellationToken);
-        public Task CreateAchievement(Achievement achievement, CancellationToken cancellationToken);
-        public Task UpdateAchievement(Achievement achievement, CancellationToken cancellationToken);
-        public Task DeleteAchievement(Guid AchievementId, CancellationToken cancellationToken);
-
-    }
-}
