--- ProjectBackEnd/Gamification/Gamification/DAL/Repositories/UnitOfWork/IUnitOfWork.cs
@@ -1,15 +1,0 @@
-﻿using Gamification.DAL.IRepository;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Gamification.DAL.Repository.UnitOfWork
-{
-    public interface IUnitOfWork
-    {
-        Task SaveChanges(CancellationToken cancellationToken);
-        IAchievementRepository achievementRepository { get; set; }
-    }
-}
