+++ ProjectBackEnd/Gamification/Gamification/Models/Achievement.cs
@@ -11,7 +11,5 @@ namespace Gamification.Models
         public string Description { get; set; }
         public int Xp { get; set; }
         public int IconId { get; set; }
-        public ICollection<UserAchievement> UserAchievements { get; set; }
-
     }
 }
