+++ ProjectBackEnd/Gamification/Gamification/Models/BaseEntity.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -9,8 +7,6 @@ namespace Gamification.Models
 {
     public class BaseEntity
     {
-        [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        [Required]
-        public Guid Id { get; set; }
+        public int Id { get; set; }
     }
 }
