--- ProjectBackEnd/Gamification/Gamification/Models/Role.cs
@@ -1,16 +1,0 @@
-﻿/*using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Gamification.Models
-{
-    public class Role : BaseEntity
-    {
-        [Required]
-        public string RoleName { get; set; }
-        public ICollection<UserRole> UserRoles { get; set; }
-    }
-}
-*/
