--- ProjectBackEnd/Gamification/Gamification/Models/UserRole.cs
@@ -1,19 +1,0 @@
-﻿/*using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Gamification.Models
-{
-    public class UserRole:BaseEntity
-    {
-        [Required]
-        public Guid UserId { get; set; }
-        public User User { get; set; }
-        [Required]
-        public Guid RoleId { get; set; }
-        public Role Role { get; set; }
-    }
-}
-*/
